+++ src/RRCFilters.jl
@@ -3,12 +3,13 @@ module RRCFilters
 using Random: AbstractRNG, default_rng, randn
 using Statistics: mean
 
-export agc, awgn, biterr, carriersync, convenc, crcconfig, crcdetect, crcgenerate, equalize, evm, freqoffset, poly2trellis, qamdemod, qammod, pskdemod, pskmod, rcosdesign, symbolsync, timingoffset, upfirdn, vitdec
+export agc, awgn, biterr, carriersync, convenc, crcconfig, crcdetect, crcgenerate, equalize, evm, freqoffset, poly2trellis, pskdemod, pskmod, psksoftdemod, qamdemod, qammod, qamsoftdemod, rcosdesign, symbolsync, timingoffset, upfirdn, vitdec
 
 include("common.jl")
 include("metrics.jl")
 include("qam.jl")
 include("psk.jl")
+include("softdemod.jl")
 include("awgn.jl")
 include("rcosdesign.jl")
 include("upfirdn.jl")